pyproject.toml
[project]
name = "BetterBio"
version = "1.0.2"
description = "A self-hosted links website based on the style of a Discord profile."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "discord-py>=2.6.4",
    "flask>=3.1.2",
]

[project.scripts]
betterbio = "src.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
